pyproject.toml
[project]
name = "veesion"
version = "0.1.0"
description = ""
authors = [
    {name = "Yohann Gabory",email = "boblefrag@gmail.com"}
]
license = {text = "GPL"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2,<6.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-rq (>=3.0.0,<4.0.0)",
    "coverage (>=7.8.0,<8.0.0)",
    "ipdb (>=0.13.13,<0.14.0)",
    "requests (>=2.32.3,<3.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "types-requests (>=2.32.0.20250328,<3.0.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylsp.plugins.mypy]
enabled = true
live_mode = true

[mypy]
ignore_missing_imports = true
